+++ testsuite/tests/rename/should_fail/T14032c.hs
@@ -1,0 +1,10 @@
+
+module T14032c where
+
+
+f $ a = f a
+
+type f $ a = f a
+
+infix 0 type $
+infix 0 data $